@@ -1,18 +1,18 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript { 
-    ext.kotlin_version = '1.5.21'
+    ext.kotlin_version = '1.7.0'
 
     repositories {
         google()
         mavenCentral()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:7.0.3'
+        classpath 'com.android.tools.build:gradle:7.0.4'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-        classpath 'com.google.dagger:hilt-android-gradle-plugin:2.28-alpha'
-        classpath 'com.google.gms:google-services:4.3.10'
-        classpath 'com.google.android.gms:oss-licenses-plugin:0.10.4'
+        classpath 'com.google.dagger:hilt-android-gradle-plugin:2.42'
+        classpath 'com.google.gms:google-services:4.3.13'
+        classpath 'com.google.android.gms:oss-licenses-plugin:0.10.5'
     }
 }
 
