@@ -1,10 +1,11 @@
 package com.github.sikv.photos.ui.adapter.viewholder
 
 import android.view.View
+import android.widget.ImageButton
+import android.widget.ImageView
+import android.widget.TextView
 import androidx.lifecycle.LifecycleCoroutineScope
 import androidx.recyclerview.widget.RecyclerView
-import com.bumptech.glide.load.resource.bitmap.CircleCrop
-import com.bumptech.glide.load.resource.drawable.DrawableTransitionOptions
 import com.github.sikv.photos.R
 import com.github.sikv.photos.manager.PhotoLoader
 import com.github.sikv.photos.model.Photo
@@ -12,8 +13,10 @@ import com.github.sikv.photos.model.PhotoItemLayoutType
 import com.github.sikv.photos.model.getAttributionPlaceholderBackgroundColor
 import com.github.sikv.photos.model.getAttributionPlaceholderTextColor
 import com.github.sikv.photos.ui.adapter.OnPhotoActionListener
-import com.github.sikv.photos.util.*
-import kotlinx.android.synthetic.main.item_photo_full.view.*
+import com.github.sikv.photos.util.OnHoldReleaseListener
+import com.github.sikv.photos.util.TextPlaceholder
+import com.github.sikv.photos.util.favoriteAnimation
+import com.github.sikv.photos.util.setOnHoldReleaseListener
 import kotlinx.coroutines.launch
 
 class PhotoViewHolder(
@@ -21,6 +24,16 @@ class PhotoViewHolder(
     private val photoLoader: PhotoLoader,
     private val lifecycleScope: LifecycleCoroutineScope
 ) : RecyclerView.ViewHolder(itemView) {
+
+    private val photographerNameText = itemView.findViewById<TextView>(R.id.photographerNameText)
+    private val sourceText = itemView.findViewById<TextView>(R.id.sourceText)
+    private val favoriteButton = itemView.findViewById<ImageButton>(R.id.favoriteButton)
+    private val photoImage = itemView.findViewById<ImageView>(R.id.photoImage)
+    private val photographerImage = itemView.findViewById<ImageView>(R.id.photographerImage)
+    private val photographerLayout = itemView.findViewById<View>(R.id.photographerLayout)
+    private val optionsButton = itemView.findViewById<ImageButton>(R.id.optionsButton)
+    private val shareButton = itemView.findViewById<ImageButton>(R.id.shareButton)
+    private val downloadButton = itemView.findViewById<ImageButton>(R.id.downloadButton)
 
     fun bind(
         itemLayoutType: PhotoItemLayoutType,
@@ -35,18 +48,20 @@ class PhotoViewHolder(
 
         loadPhotos(photo, lifecycleScope)
 
-        itemView.photographerNameText.text = photo?.getPhotoPhotographerName()
-        itemView.sourceText.text = photo?.getPhotoSource()?.title
+        photographerNameText.text = photo?.getPhotoPhotographerName()
+        sourceText.text = photo?.getPhotoSource()?.title
 
-        itemView.favoriteButton.visibility = View.VISIBLE
-        itemView.favoriteButton.setImageResource(if (isFavorite) R.drawable.ic_favorite_red_24dp else R.drawable.ic_favorite_border_24dp)
+        favoriteButton.visibility = View.VISIBLE
+        favoriteButton.setImageResource(
+            if (isFavorite) R.drawable.ic_favorite_red_24dp else R.drawable.ic_favorite_border_24dp
+        )
 
         photo?.let {
-            itemView.photoImage.setOnClickListener { view ->
+            photoImage.setOnClickListener { view ->
                 listener.onPhotoAction(OnPhotoActionListener.Action.CLICK, photo, view)
             }
 
-            itemView.photoImage.setOnHoldReleaseListener(object : OnHoldReleaseListener() {
+            photoImage.setOnHoldReleaseListener(object : OnHoldReleaseListener() {
                 override fun onHold(view: View) {
                     listener.onPhotoAction(OnPhotoActionListener.Action.HOLD, photo, view)
                 }
@@ -56,53 +71,53 @@ class PhotoViewHolder(
                 }
             })
 
-            itemView.photographerImage.setOnClickListener { view ->
+            photographerImage.setOnClickListener { view ->
                 listener.onPhotoAction(OnPhotoActionListener.Action.PHOTOGRAPHER, photo, view)
             }
 
-            itemView.photographerLayout.setOnClickListener { view ->
+            photographerLayout.setOnClickListener { view ->
                 listener.onPhotoAction(OnPhotoActionListener.Action.PHOTOGRAPHER, photo, view)
             }
 
-            itemView.optionsButton.setOnClickListener { view ->
+            optionsButton.setOnClickListener { view ->
                 listener.onPhotoAction(OnPhotoActionListener.Action.OPTIONS, photo, view)
             }
 
-            itemView.favoriteButton.setOnClickListener { view ->
+            favoriteButton.setOnClickListener { view ->
                 listener.onPhotoAction(OnPhotoActionListener.Action.FAVORITE, photo, view)
-                itemView.favoriteButton.favoriteAnimation()
+                favoriteButton.favoriteAnimation()
             }
 
-            itemView.shareButton.setOnClickListener { view ->
+            shareButton.setOnClickListener { view ->
                 listener.onPhotoAction(OnPhotoActionListener.Action.SHARE, photo, view)
             }
 
-            itemView.downloadButton.setOnClickListener { view ->
+            downloadButton.setOnClickListener { view ->
                 listener.onPhotoAction(OnPhotoActionListener.Action.DOWNLOAD, photo, view)
             }
 
         } ?: run {
-            itemView.optionsButton.setOnClickListener(null)
-            itemView.photoImage.setOnClickListener(null)
-            itemView.photoImage.setOnHoldReleaseListener(null)
-            itemView.favoriteButton.setOnClickListener(null)
-            itemView.downloadButton.setOnClickListener(null)
+            optionsButton.setOnClickListener(null)
+            photoImage.setOnClickListener(null)
+            photoImage.setOnHoldReleaseListener(null)
+            favoriteButton.setOnClickListener(null)
+            downloadButton.setOnClickListener(null)
         }
     }
 
     private fun bindMin(photo: Photo?, listener: OnPhotoActionListener) {
-        itemView.photoImage.setImageDrawable(null)
+        photoImage.setImageDrawable(null)
 
         photo?.let {
-            photoLoader.load(it.getPhotoPreviewUrl(), itemView.photoImage)
+            photoLoader.load(it.getPhotoPreviewUrl(), photoImage)
         }
 
         photo?.let {
-            itemView.photoImage.setOnClickListener { view ->
+            photoImage.setOnClickListener { view ->
                 listener.onPhotoAction(OnPhotoActionListener.Action.CLICK, photo, view)
             }
 
-            itemView.photoImage.setOnHoldReleaseListener(object : OnHoldReleaseListener() {
+            photoImage.setOnHoldReleaseListener(object : OnHoldReleaseListener() {
                 override fun onHold(view: View) {
                     listener.onPhotoAction(OnPhotoActionListener.Action.HOLD, photo, view)
                 }
@@ -113,20 +128,20 @@ class PhotoViewHolder(
             })
 
         } ?: run {
-            itemView.photoImage.setOnClickListener(null)
-            itemView.photoImage.setOnHoldReleaseListener(null)
+            photoImage.setOnClickListener(null)
+            photoImage.setOnHoldReleaseListener(null)
         }
     }
 
     private fun loadPhotos(photo: Photo?, lifecycleScope: LifecycleCoroutineScope) {
-        itemView.photoImage.setImageDrawable(null)
-        itemView.photographerImage.setImageDrawable(null)
+        photoImage.setImageDrawable(null)
+        photographerImage.setImageDrawable(null)
 
         if (photo == null) {
             return
         }
 
-        photoLoader.load(photo.getPhotoPreviewUrl(), itemView.photoImage)
+        photoLoader.load(photo.getPhotoPreviewUrl(), photoImage)
 
         val textColor = photo.getAttributionPlaceholderTextColor(itemView.context)
         val backgroundColor = photo.getAttributionPlaceholderBackgroundColor(itemView.context)
@@ -141,7 +156,7 @@ class PhotoViewHolder(
             photoLoader.loadWithCircleCrop(
                 photo.getPhotoPhotographerImageUrl(),
                 placeholder,
-                itemView.photographerImage
+                photographerImage
             )
         }
     }
